+++ src/array/array_sync_sharded_readable_ext.rs
@@ -9,7 +9,7 @@ use super::{
     codec::CodecOptions, concurrency::concurrency_chunks_and_codec, Array, ArrayError,
     ArrayShardedExt, ChunkGrid, UnsafeCellSlice,
 };
-use super::{ArrayBytes, DataTypeSize};
+use super::{ArrayBytes, ArraySize, DataTypeSize};
 use crate::storage::ReadableStorageTraits;
 use crate::{array::codec::ArrayPartialDecoderTraits, array_subset::ArraySubset};
 
@@ -340,10 +340,9 @@ impl<TStorage: ?Sized + ReadableStorageTraits + 'static> ArrayShardedReadableExt
             // Retrieve chunk bytes
             let num_shards = shards.num_elements_usize();
             if num_shards == 0 {
-                Ok(ArrayBytes::new_fill_value(
-                    array_subset.num_elements_usize(),
-                    self.fill_value(),
-                ))
+                let array_size =
+                    ArraySize::new(self.data_type().size(), array_subset.num_elements());
+                Ok(ArrayBytes::new_fill_value(array_size, self.fill_value()))
             } else {
                 // Calculate chunk/codec concurrency
                 let chunk_representation =