+++ src/array/codec/array_to_bytes/sharding/sharding_codec.rs
@@ -12,8 +12,8 @@ use crate::{
         concurrency::calc_concurrency_outer_inner,
         transmute_to_bytes_vec, unravel_index,
         unsafe_cell_slice::UnsafeCellSlice,
-        ArrayBytes, ArrayMetadataOptions, BytesRepresentation, ChunkRepresentation, ChunkShape,
-        DataTypeSize, FillValue, RawBytes,
+        ArrayBytes, ArrayMetadataOptions, ArraySize, BytesRepresentation, ChunkRepresentation,
+        ChunkShape, DataTypeSize, FillValue, RawBytes,
     },
     array_subset::ArraySubset,
     metadata::v3::MetadataV3,
@@ -221,10 +221,11 @@ impl ArrayToBytesCodecTraits for ShardingCodec {
                     let offset = shard_index[chunk_index * 2];
                     let size = shard_index[chunk_index * 2 + 1];
                     let chunk_bytes = if offset == u64::MAX && size == u64::MAX {
-                        ArrayBytes::new_fill_value(
-                            chunk_representation.num_elements_usize(),
-                            chunk_representation.fill_value(),
-                        )
+                        let array_size = ArraySize::new(
+                            chunk_representation.data_type().size(),
+                            chunk_representation.num_elements(),
+                        );
+                        ArrayBytes::new_fill_value(array_size, chunk_representation.fill_value())
                     } else if usize::try_from(offset + size).unwrap() > encoded_shard.len() {
                         return Err(CodecError::Other(
                             "The shard index references out-of-bounds bytes. The chunk may be corrupted."