+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `DataType::fixed_size()` that returns `Option<usize>`. Returns `None` for variable length data types.
  - Add `ArrayError::IncompatibleElementType` (replaces `ArrayError::IncompatibleElementSize`)
  - Add `ArrayError::InvalidElementValue`
+ - Add `ChunkShape::num_elements_u64`
 
 ### Changed
  - Use `[async_]retrieve_array_subset_opt` internally in `Array::[async_]retrieve_chunks_opt`